Dice roll simulation
This is a web app where every time you refresh the page, it will roll two dice. It uses a random number generator from 1-6 to simulate the six sides of the dice. If the first dice is higher, player 1 wins. if the second dice is higher, player 2 wins. If they are the same, it is a draw