Drum sounds with keyboard press
This web app will have pictures displayed on the screen of the different parts of a drum set. They will have letters above them to show that you can press the corresponding key on the keyboard to have the sound play. Or you can click on the image to play the sound as well.