Mongo DB Demonstration
This demonstrates how to create a database using mongo DB and to specify the fields that are needed to insert data. Once the data is correct, it will run an Insert statement to insert what the user has written into the database. If they change anything, an Update function will trigger to change what needs to be changed.