Planner
This is a React web app where the user is able to add a title and content to a little tag. It will display all of the things the user has posted. It can be a todo list, a reminder list, or anything you want. It can be equivalent to sticky notes. Once the task is done, there is a delete button to get rid of the accomplished task.