ToDo list with React
This is a simple todo list using React. The user will be able to input what they want onto the list and have it added below the test area. They can add multiple items. Once you have finished that item, you can click on it and it will be deleted. They are organized by bullet points.