Mapping Components with React
This is a simple web app using React that gives you the meaning of what emoji's mean. The creator is able to add however many he wants as long as the emoji, name, and description fields are filled in. This uses multiple React files to work. This way, the work is being split up into multiple files that serve a single purpose instead of having all of it in one spot. It also helps with problem solving.