Simon game
This is a recreation of the old school Simon game. There are four buttons on screen that are red, yellow, green, and blue. This uses jQuery to have the game function as well as keep adding to the array of the pattern with a random number generator selecting a color at random. When the user tries to follow along, their choices will be stored into a separate array to see if they match with the pattern the game made.