Simple ToDo list
This is a simple todo list involving EJS. There is a text box that is there for the user to be able to create things that are needed for the list. There is also a different page for work as well as one for home life. That way you don't have to have everything cluttered together.