Wiki page with Database
This is a type of Wiki web app that uses a database to store the data that has been passed in. There is the ability to add more articles with a title and content and have them save to the database and have them be displayed. There is also the option to delete the articles from the database.